S12P21A303-70-feat: 하나의 기부 정보를 읽어 오는 DTO, Controller 생성

+++ Back/src/main/java/com/ssafy/Dandelion/domain/autodonation/entity/AutoDonationInfo.java
@@ -29,7 +29,7 @@ public class AutoDonationInfo extends BaseTimeEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long autoDonationInfoId;
+	private Integer autoDonationInfoId;
 
 	@Column(nullable = false)
 	private Integer autoDonationId;
